+++ nyxx_lavalink/lib/src/model/WebSocketClosed.dart
@@ -2,8 +2,6 @@ part of nyxx_lavalink;
 
 /// Web socket closed event from lavalink
 class WebSocketClosedEvent extends BaseEvent {
-  /// Type of close
-  late final String? type;
   /// Guild where the websocket has closed
   late final Snowflake guildId;
   /// Close code
@@ -15,7 +13,6 @@ class WebSocketClosedEvent extends BaseEvent {
 
   WebSocketClosedEvent._fromJson(Nyxx client, Node node, Map<String, dynamic> json) : super(client, node) {
     this.guildId = Snowflake(json["guildId"]);
-    this.type = json["type"] as String?;
     this.code = json["code"] as int;
     this.reason = json["reason"] as String;
     this.byRemote = json["byRemote"] as bool;